requirements.txt
'''
This file lists the Python packages that your project depends on. You should list all of your project's dependencies in this file.
'''
discord.py
requests